split main.py
